# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";
import type { ChatRequestAnswerMode } from "./chatRequestAnswerMode";

export interface ChatRequest {
  /** User's Gemini API key */
  apiKey: string;
  messages: ChatMessage[];
  /** Answer mode for AI response style */
  answerMode: ChatRequestAnswerMode;
  /** Optional system instruction override */
  systemInstruction?: string;
}