# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatResponseUsage } from "./chatResponseUsage";

export interface ChatResponse {
  reply: string;
  usage?: ChatResponseUsage;
}