# lib/api-zod/src/generated/types/inviteInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";

export interface InviteInfo {
  token: string;
  subjectName: string;
  sessionTitle: string;
  sessionId: string;
  messages?: ChatMessage[];
  createdBy?: string;
  createdAt: string;
}